S09P31A402-186 [FEATURE] refresh token 갱신 기능 추가

+++ backend/auth/src/main/java/com/auth/domain/member/entity/Member.java
@@ -15,7 +15,7 @@ public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(columnDefinition = "INT UNSIGNED")
-    private Integer memberId;
+    private Long memberId;
 
     @Column(nullable = false, unique = true, length = 320)
     private String email;
